proto: move decode_public_key() implementation
